* master: (57 commits)
  Readme
  Readme
  Readme
  Accept -T as a short command line flag for --tmout_sigvtalrm
  Reset fuzzer->tmOutSignaled when starting a new subprocess
  Kill the process with SIGKILL is the first signal did not cause the process to exit
  Add support for SIGVTALRM to MacOS code
  cmdline: allow to kill process with SIGVTALRM is it timeouts
  linux: fix arch_perfInit if Intel PT is not available
  persistent: ssize_t instead of size_t for reads
  persistent: accept input from stdin if run separately
  posix: check time limit during interrupted calls only
  posix/arch: check for time-limits less frequently
  linux: don't specify SA_SIGINFO in sighandlers as it's not needed
  Modify the persistent socket only during execve
  Readme: Correct CVE
  No need to stop process in persistent mode
  Revert "persistent: signal ready from persistent process"
  persistent: signal ready from persistent process
  Minimal undefined file size is 128 bytes
  ...

+++ subproc.c
@@ -124,6 +124,32 @@ const char *subproc_StatusToStr(int status, char *str, size_t len)
     return str;
 }
 
+bool subproc_persistentModeRoundDone(honggfuzz_t * hfuzz, fuzzer_t * fuzzer)
+{
+    if (hfuzz->persistent == false) {
+        return false;
+    }
+    char z;
+    if (recv(fuzzer->persistentSock, &z, sizeof(z), MSG_DONTWAIT) == sizeof(z)) {
+        LOG_D("Persistent mode round finished");
+        return true;
+    }
+    return false;
+}
+
+static bool subproc_persistentSendFile(fuzzer_t * fuzzer)
+{
+    uint32_t len = (uint64_t) fuzzer->dynamicFileSz;
+    if (files_writeToFd(fuzzer->persistentSock, (uint8_t *) & len, sizeof(len)) == false) {
+        return false;
+    }
+    if (files_writeToFd(fuzzer->persistentSock, fuzzer->dynamicFile, fuzzer->dynamicFileSz) ==
+        false) {
+        return false;
+    }
+    return true;
+}
+
 bool subproc_PrepareExecv(honggfuzz_t * hfuzz, fuzzer_t * fuzzer, const char *fileName)
 {
     /*
@@ -178,12 +204,13 @@ bool subproc_PrepareExecv(honggfuzz_t * hfuzz, fuzzer_t * fuzzer, const char *fi
     return true;
 }
 
-bool subproc_New(honggfuzz_t * hfuzz, fuzzer_t * fuzzer)
+static bool subproc_New(honggfuzz_t * hfuzz, fuzzer_t * fuzzer)
 {
     fuzzer->pid = fuzzer->persistentPid;
     if (fuzzer->pid != 0) {
         return true;
     }
+    fuzzer->tmOutSignaled = false;
 
     int sv[2];
     if (hfuzz->persistent) {
@@ -239,39 +266,20 @@ bool subproc_New(honggfuzz_t * hfuzz, fuzzer_t * fuzzer)
     return true;
 }
 
-bool subproc_persistentModeRoundDone(honggfuzz_t * hfuzz, fuzzer_t * fuzzer)
+bool subproc_Run(honggfuzz_t * hfuzz, fuzzer_t * fuzzer)
 {
-    if (hfuzz->persistent == false) {
+    if (subproc_New(hfuzz, fuzzer) == false) {
+        LOG_E("subproc_New()");
         return false;
     }
-    char z;
-    if (recv(fuzzer->persistentSock, &z, sizeof(z), MSG_DONTWAIT) == sizeof(z)) {
-        LOG_D("Persistent mode round finished");
-        return true;
-    }
-    return false;
-}
 
-static bool subproc_persistentSendFile(fuzzer_t * fuzzer)
-{
-    uint32_t len = (uint64_t) fuzzer->dynamicFileSz;
-    if (files_writeToFd(fuzzer->persistentSock, (uint8_t *) & len, sizeof(len)) == false) {
-        return false;
-    }
-    if (files_writeToFd(fuzzer->persistentSock, fuzzer->dynamicFile, fuzzer->dynamicFileSz) ==
-        false) {
-        return false;
-    }
-    return true;
-}
-
-void subproc_Run(honggfuzz_t * hfuzz, fuzzer_t * fuzzer)
-{
     arch_prepareChild(hfuzz, fuzzer);
     if (hfuzz->persistent == true && subproc_persistentSendFile(fuzzer) == false) {
         LOG_W("Could not send file contents to the persistent process");
     }
     arch_reapChild(hfuzz, fuzzer);
+
+    return true;
 }
 
 uint8_t subproc_System(const char *const argv[])
@@ -322,9 +330,21 @@ void subproc_checkTimeLimit(honggfuzz_t * hfuzz, fuzzer_t * fuzzer)
     int64_t curMillis = util_timeNowMillis();
     int64_t diffMillis = curMillis - fuzzer->timeStartedMillis;
     if (diffMillis > (hfuzz->tmOut * 1000)) {
-        LOG_W("PID %d took too much time (limit %ld s). Sending SIGKILL",
-              fuzzer->pid, hfuzz->tmOut);
-        kill(fuzzer->pid, SIGKILL);
+        /* Has this instance been already signaled due to timeout? Just, SIGKILL it */
+        if (fuzzer->tmOutSignaled) {
+            LOG_W("PID %d has already been signaled due to timeout. Killing it with SIGKILL",
+                  fuzzer->pid);
+            kill(fuzzer->pid, SIGKILL);
+            return;
+        }
+        fuzzer->tmOutSignaled = true;
+
+        LOG_W("PID %d took too much time (limit %ld s). Killing it", fuzzer->pid, hfuzz->tmOut);
+        if (hfuzz->tmout_vtalrm) {
+            kill(fuzzer->pid, SIGVTALRM);
+        } else {
+            kill(fuzzer->pid, SIGKILL);
+        }
         ATOMIC_POST_INC(hfuzz->timeoutedCnt);
     }
 }
